REVIEW ME Avoid overflow from 64-bit `A * B / C` in `std::sys::windows::time`.
I should check if this method for "safer" multiplication by a ratio is
already provided somewhere in `core::num`.  (And if not, add it.)

Also, I should look into whether I have any better options here, and
whether I should just be calling the WrappingOps or OverflowingOps
anyway. But this should serve for the short term.